Fix meat trigger.
Respawn the meat if it goes under the map.
Convert 0s, amd 1s to booleans if applicable. Add prototype function to check if a point is inside a 2d polygon(will be used for making the meat bounce off the polygon instead of going outside the map). Prevent perks and packapunch from spawning in.